Add ProofNode: alternative Merkle proof structure
With this design, the three cases (present, in empty branch, or in
branch with mismatched leaf) aren't explicitly split out. Instead,
the server just returns enough nodes to the client that the client
can simulate the lookup itself, verifying that it gets the same leaf
and the hashes line up.