Implement promise storage using etcd
I expect this code to be run on the frontend servers. I had to
base64-encode the serialized protobufs for things to work (Certificate
Transparency does the same thing); I guess etcd's HTTP API is a bit
fragile. I've also included a simple test client (needs to be pointed at
an etcd cluster).