Add a ConsistentStorage interface
This follows the new design. I tried to make sure that the
implementation of this interface with etcd would be fairly simple and
building both the signer and the frontend nodes on top of the interface
would be doable. I kept the interface relatively agnostic to choices
like how soon we're going to promise updates will make it into the tree.
Mostly, the interface is modeled after consistent_store.h in the
open-source CT server, except with some additional logic to handle
promises.