Multi-tenant configuration support
Support multiple backend logs by storing their configuration
in an in-memory table.

In the future, the Add* commands of the Admin interface will perform
the needed setup such as inserting master rows in the various
sub-components.

Further in the future, this in-memory table will be backed by a database
rather than configured in-memory from the commandline.