Use log tree size as current revision
Fixes #753

By using the log as the authoritative source for the latest revision, we
avoid the scenario where by the map has produced a new revision which
has not been sequenced (yet) by the log.

During this window of a time, a request for the latest revision produces
a log inclusion proof error.