Use raw trillian.Log connection in Signer
Previously, the signer used a verifying trillian log client in the
signer. This:
- Required a separate build step to acquire the verifying key
- Was tied to a confusing and only partially implemented multi-tennant
  configuration system.

This commit simplifies the situation to make a pure grpc call to the
trillian log backend, simplifying debugging, and makes things consistent
consistent with the way we are querying the trillian map server, and is
also consistent with the way the Certificate Transparency frontends
treat their own trillian backends.