Move KeyValue into SignedKV
Remove one level of proto encapsulation by embedding the signature in
SignedKV along with the data to be signed.

Before ObjectHash, protos needed to be serialized then signed, forcing
us to create two proto messages for each signed type: a value type and a
signed value type. With ObjectHash we can avoid signing a particular
serialization of a protocol buffer and just sign the object itself,
inserting the signature into the object after signing.  Verification
involves removing the signature before verifying the object.

This is part of the proto cleanup effort in  XXX