Simplify mutation protos  (#823)
* Move KeyValue into SignedKV

Remove one level of proto encapsulation by embedding the signature in
SignedKV along with the data to be signed.

Before ObjectHash, protos needed to be serialized then signed, forcing
us to create two proto messages for each signed type: a value type and a
signed value type. With ObjectHash we can avoid signing a particular
serialization of a protocol buffer and just sign the object itself,
inserting the signature into the object after signing.  Verification
involves removing the signature before verifying the object.

This is part of the proto cleanup effort in  XXX

* EntryUpdate.Update -> Mutation

* Mutation.Value type []byte -> Entry

It appears that grpc has trouble with proto3 OneOf.
We didn't have problems before when entry was a serialized []byte,
but now that we're passing a regular proto object, the PublicKey type
is coming out the otherside, stripped of it's interior values.
This has been fixed by #821

* cleanup

* Remove spurious gogo dependency