Generate protos from a common directory
protoc uses the path passed to protoc as the prefix for symbol
resolution. If protoc is invoked from different directories, the
resulting FileDescriptors will have different symbols, either causing
the symbols to duplicate, or be unresolvable.

The solution in this commit is to generate all the protos for this
project from the same directory like we did for Trillian.

This also has the side benefit of producing a single directory for
reverse symbol lookkup when using grpc reflection with the grpc_cli
tool.