Fix upstream trillian db changes
Trillian recently switched from MySQL to SQLite as it's default database
for testing. This is great, except that SQLite doesn't support
concurrent writes, something that the Trillian Map makes extensive use
of.

This PR adds the needed flags to run Key Transparency with a MySQL db
backend. When a MySQL is not available, it skips the intergration tests.