Remove transactions from mutation api
Transactions are not currently serving any purpose.
There are no sequences of database calls that need to be coordinated
atomically.

The transaction abstraction was orginally created to support atomically
acking an item out of a queue and saving it.
1. The queue and the database can often be using different technologies
that make a single transaction between them not possible.
2. The specifics of the queue, databases, and other subsystems are
implementation details that should be handeled in impl/ rather than
creeping into the apis themselves.