Split Docker image building from running.
If a `build` stanza is present in the docker-compose.yml file,
docker-compose pull will not run if the build stanza's paths do
not exist. This prevents consumers from simply pulling the docker
images from the us.gcr repository.

This allows customers to use docker-compose pull to run KeyTransparency
without downloaded the rest of the source code.