Blocking AddLeaf in Sequencer
The Trillian Log API contract does NOT guaruntee that items will be
sequenced in the same order they are submitted.  Eventually we will use
the AddSequencedLeavesAPI, but until that is available, we can refuse to
send any more leaves to the log until it as sequenced the one we just
sent it.