Move VerifyEpoch to Client (#966)
* Move VerifyEpoch to Client

We want to encourage a new API pattern of:

1. Get current epoch
2. Query the API at that epoch.

As a result, getting and verifying epochs is functionality that is
common to all clients. This move removes some duplication and makes the
API simpler to understand.

* Remove unused trusted reference in Monitor

The trusted anchor is provided by main().
TODO: Save last seen SignedLogRoot to disk.
TODO: Gossip SignedLogRoot.

* Use VerifyEpoch in GetEntryResponse

* Verify latest epoch matches log.TreeSize in GetLatestEpoch

* remove duplicate statement

* Move treesize printfs to updateroot

* Reviewer comments

* Acquire trustedLock

* Acquire trusted lock in inner loop of listhistory

* Break VerifiedListHistory into a separate function

This commit breaks out the inner loop of client.ListHistory into
a separate function which only makes a single verified API call.

It also breaks out the compressing of duplicate items in the list of
profiles to a separate, testable function.