Clean up GetEntry API as GetUser
Epoch already defines the grouping of fields needed to verify a signed map head within a log. We can define a sibling message to group the fields needed to verify a leaf within a signed map head.

This will allow us to define a BatchGetUser API later on (#588) which returns a single Epoch and multiple MapLeafEntries.