New HighWatermark tests that respect interface.
Replace tests that wanted to predict the exact value of HighWatermark
with tests that rely on the behavior of ReadLog to assert that the value
returned is indeed correct.

This allows different storage layers to use their own strategies
for returning high watermarks.

It also removes the requirement that HighWatermark values themselves
be stored with nanosecond precision.