Query logs by watermark instead of Time (#1404)
The reason for having time.Time in the storage API is primarily monitoring.
Watermarks are currently used to compute latency of different stages of the
pipeline.

There are many disadvantages though:

It hurts to review PRs that operate with time.Time magic in queries.
Zero value needs special treatment.
There are 3(!) different types for time: Golang's time.Time,
protobuf's Timestamp, and MySQL DATETIME. We depend on conversions
between these types to work correctly.
Storage implementations need to do arcane rounding arithmetics with
the timestamp. It feels like using float instead of int.
Using time.Time in the API prevents non-Time implementations, e.g.
simple sequence numbers like in Kafka. For the problem that these
timestamps are solving, it actually makes more sense to use
non-timestamp types.
The proposed solution is to use obscure water.Mark type for queries, and
make timestamp a field of the entry type to support the metrics case. Since
metrics don't need to be as precise as query watermarks, this will massively
simplify the code. This PR removes time.Time semantics, and follow-up PRs
will reintroduce time.Time in a form of a field.